+++ java/nets/src/alfred/Net.java
@@ -1,6 +1,7 @@
 package alfred;
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -14,7 +15,7 @@ public abstract class Net {
 	 * 
 	 * @return output of this Net.
 	 */
-	public abstract double getOutput();
+	public abstract BigDecimal getOutput();
 	
 	public enum NetType {
 		TRAIN("augtrain"),
